+++ packages/ui/src/components/date-picker/date-picker.component.tsx
@@ -7,6 +7,7 @@ import { type DatePickerProps, DuetDatePickerElement } from './date-picker.types
 import { formatDate, isDateDisabled, useListener } from './date-picker.utils.js';
 
 export function DatePicker({
+  dateFormat = 'dd-MM-yyyy',
   disableWeekends,
   disableDaysOfWeek,
   disableDates,
@@ -23,6 +24,7 @@ export function DatePicker({
   size = 'md',
   name,
   block = false,
+  invalid = false,
   ...props
 }: DatePickerProps) {
   const [initialized, setInitialized] = useState(false);
@@ -47,16 +49,16 @@ export function DatePicker({
   const dateAdapter = useMemo(
     () => ({
       parse(value = '', createDate: (year: string, month: string, day: string) => Date) {
-        const matches = value.match(/^(\d{1,2})-(\d{1,2})-(\d{4})$/);
+        const matches = value.match(/^(\d{1,2})[-/](\d{1,2})[-/](\d{4})$/);
         if (matches) {
           return createDate(matches[3], matches[2], matches[1]);
         }
       },
       format(date: Date) {
-        return formatDate(date, 'dd-mm-yyyy');
+        return formatDate(date, dateFormat as 'dd-MM-yyyy' | 'dd/MM/yyyy');
       },
     }),
-    [],
+    [dateFormat],
   );
 
   const localization = useMemo(() => {
@@ -122,5 +124,5 @@ export function DatePicker({
 
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
-  return <duet-date-picker class={styles({ size, block })} ref={ref} {...props} />;
+  return <duet-date-picker class={styles({ size, block, invalid })} ref={ref} {...props} />;
 }