+++ apps/protoform/CHANGELOG.md
@@ -1,5 +1,75 @@
 # protoform
 
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [b31459e]
+  - @westpac/ui@0.40.3
+
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies [18c66ea]
+  - @westpac/ui@0.40.2
+
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies [6b3ed1b]
+  - @westpac/ui@0.40.1
+
+## 0.2.19
+
+### Patch Changes
+
+- Updated dependencies [04d5a45]
+  - @westpac/ui@0.40.0
+
+## 0.2.18
+
+### Patch Changes
+
+- Updated dependencies [faea8ea]
+  - @westpac/ui@0.39.0
+
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies [779e98f]
+  - @westpac/ui@0.38.0
+
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies [66b74a9]
+  - @westpac/ui@0.37.0
+
+## 0.2.15
+
+### Patch Changes
+
+- Updated dependencies [3f410a7]
+  - @westpac/ui@0.36.1
+
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies [40d387d]
+  - @westpac/ui@0.36.0
+
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [2ed1fd5]
+  - @westpac/ui@0.35.0
+
 ## 0.2.12
 
 ### Patch Changes